+++ kernel/src/syscall/mod.rs
@@ -637,8 +637,6 @@ impl Syscall {
                 )
             }
             #[cfg(target_arch = "x86_64")]
-            SYS_UTIMES => Self::sys_utimes(args[0] as *const u8, args[1] as *const PosixTimeval),
-            #[cfg(target_arch = "x86_64")]
             SYS_EVENTFD => {
                 let initval = args[0] as u32;
                 Self::sys_eventfd(initval, 0)
